Update (#2) createstudent.php
Added two more scripts for this code! One is a javascript that changes the original style of header1, which means it no longer uses the style from the given CSS. The 2nd is another script using Ajax to notify users already existing names when they try to create a new student.