Update (#3) changestudent.php
Simply added another script for changing the styles of header1 for this file similarly as createstudent.php. For the Ajax code, I am wondering if I should implement one in this file for permanently deleting the student on the delete student form just by adding another script code to delete the info or implement it on the deletestudent file.